% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAIC.R
\name{BAIC}
\alias{BAIC}
\title{BAIC metric function}
\usage{
BAIC(metrics, metric_values, props, fun = NULL)
}
\arguments{
\item{metrics:}{Metric names}

\item{metric_values:}{Metric values in a list to calculate the BAIC value}

\item{props:}{Proportion of minority class of each dataset to get the BAIC value}

\item{fun:}{Function to obtain the weights of each metric}
}
\value{
number
}
\description{
BAIC metric function
}
