% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILM.R
\name{FILM}
\alias{FILM}
\title{FILM function for doing the entire methodology and get the best model accordin}
\usage{
FILM(
  dataset,
  formula,
  df_aux = 6,
  IAAs = c("IPIP", "SMOTE", "ROSE"),
  models = c("ranger", "glm"),
  hyperparameters = NULL,
  metrics = c("ACCURACY", "KAPPA", "F1"),
  cv = 5,
  prop_min = 0.05,
  prop_max = 0.4,
  tC = trainControl(summaryFunction = FILM::metrics, allowParallel = TRUE, classProbs =
    TRUE),
  metric_max = "KAPPA",
  num.trees = 200,
  return.uic = T
)
}
\arguments{
\item{dataset:}{Dataset of the binary classification problem}

\item{formula:}{Style: 'target_variable ~ .'}

\item{df_aux:}{Number of auxiliar datasets with different minority class function}

\item{IAAs:}{Imbalance Aware Approaches such as 'SMOTE', 'IPIP', 'ROSE', 'Upsample' and 'Downsample'}

\item{models:}{ML techniques to be trained after using some IAA: 'glm' or 'ranger'.}

\item{hyperparameters:}{Hiperparameters for 'ranger' method.}

\item{metrics:}{Vector of metrics to get in order to calculate UIC metric. Possible values: 'ACCURACY' ,'SENS', 'SPEC', 'PPV', 'NPV', 'KAPPA' , 'BAL_ACC',  'F1', 'AUC', 'PR', 'MCC', 'GEOM'}

\item{cv:}{Number K of folds for a K-Folds Cross Validation.}

\item{prop_min:}{Minimum of proportion of the minority class for the resample auxiliar datasets.}

\item{prop_max:}{Maximum of proportion of the minority class for the resample auxiliar datasets.}

\item{tC:}{Train Control of caret train() function.}

\item{metric_max:}{Metric from 'metrics' to maximize when training ML techniques.}

\item{num.trees:}{Number of trees of the 'ranger' caret method.}
}
\value{
UIC values, best model, all models
}
\description{
FILM function for doing the entire methodology and get the best model accordin
}
