% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreementDisagreement.R
\name{integrated_agree_disagree}
\alias{integrated_agree_disagree}
\title{Integrated agreement/disagreement plot}
\usage{
integrated_agree_disagree(metrics, model = c("RF"), legend_pos = "bottom")
}
\arguments{
\item{metrics:}{Metrics from film$all_metrics}

\item{model:}{the model (RF/RLOG)}

\item{legend_pos:}{Legend position (by default "bottom")}
}
\value{
Agree and Disagree plot
}
\description{
Plot or compute agreement and disagreement between models
}
