% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPIP.R
\name{IPIP}
\alias{IPIP}
\title{IPIP FUNCTION}
\usage{
IPIP(
  formula,
  dataset,
  val = NULL,
  pmin = 0.45,
  bs = NULL,
  bE = NULL,
  model = "glm",
  mt = NULL,
  tC = trainControl(summaryFunction = FILM::metrics, classProbs = TRUE),
  hiperparametros = NULL,
  metric_max = "KAPPA",
  num.trees = 200
)
}
\arguments{
\item{num.trees}{}
}
\description{
IPIP FUNCTION
}
