% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics}
\usage{
metrics(data, lev = levels(as.factor(data$obs)), model = NULL)
}
\arguments{
\item{model}{}

\item{data:}{Data frame with the original values in column 'obs', predictions in 'pred' column}
}
\value{
Metric values for the given predictions and original values.
}
\description{
Metrics
}
