% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_dfs.R
\name{metrics_dfs}
\alias{metrics_dfs}
\title{Function to get the metric for all dfs created on FILM methodology.}
\usage{
metrics_dfs(models, formula, model_names, val, metrics)
}
\arguments{
\item{models:}{List of models}

\item{formula:}{Formula to test. Style: 'target_variable ~ .'}

\item{model_names:}{Model names to identify if it is an IPIP method.}

\item{val:}{List of validation datasets.}

\item{metrics:}{Vector of the wanted metric values.}
}
\value{
Named vector of metric values.
}
\description{
Function to get the metric for all dfs created on FILM methodology.
}
