% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIC.R
\name{UIC}
\alias{UIC}
\title{UIC metric function}
\usage{
UIC(metrics, metric_values, props, fun = NULL)
}
\arguments{
\item{metrics:}{Metric names}

\item{metric_values:}{Metric values in a list to calculate the UIC value}

\item{props:}{Proportion of minority class of each dataset to get the UIC value}

\item{fun:}{Function to obtain the weights of each metric}
}
\value{
number
}
\description{
UIC metric function
}
