% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPIP.R
\name{IPIP}
\alias{IPIP}
\title{Title}
\usage{
IPIP(
  formula,
  dataset,
  val = NULL,
  pmin = 0.45,
  bs = ceiling(log(0.01)/(log(1 - 1/nrow(dataset[which(dataset[, class] ==
    levels(dataset[, class])[which.min(dataset[, class])]), ])) *
    round(nrow(dataset[which(dataset[, class] == levels(dataset[,
    class])[which.min(dataset[, class])]), ]) * 0.75))),
  bE = ceiling(log(0.01)/(log(1 - 1/round(nrow(dataset[which(dataset[, class] ==
    levels(dataset[, class])[which.min(dataset[, class])]), ]) * 0.75)) *
    round(nrow(dataset[which(dataset[, class] == levels(dataset[,
    class])[which.min(dataset[, class])]), ]) * 0.75))),
  model = "glm",
  mt = function(n) {
     ceiling((bE - n)/3)
 },
  tC = trainControl(summaryFunction = FILM::metrics, classProbs = TRUE),
  hiperparametros = NULL,
  metric_max = "KAPPA",
  num.trees = 200
)
}
\arguments{
\item{num.trees}{}
}
\description{
Title
}
